# api/core-api/app/Http/Controllers/AuthController.php
<?php

namespace App\Http\Controllers;

use App\Models\User;
use Illuminate\Http\Request;
use App\Http\Requests\Authentication\SignInRequest;
use App\Http\Requests\Authentication\SignUpRequest;
use Illuminate\Auth\Events\Verified;
use Illuminate\Routing\Controllers\HasMiddleware;
use Illuminate\Routing\Controllers\Middleware;
use Illuminate\Support\Facades\Hash;

class AuthController extends Controller implements HasMiddleware
{
    public static function middleware()
    {
        return [
            new Middleware('auth:sanctum', only: [
                'signOut',
                'sendEmailVerification',
            ]),
            new Middleware('throttle:6,1', only: [
                'sendEmailVerification'
            ]),
            new Middleware('signed', only: [
                'verifyEmail'
            ])
        ];
    }

    /**
     * @OA\Post(
     *     path="/api/auth/email/resend-verification",
     *     tags={"Authentication"},
     *     summary="User resend email verification",
     *     description="Send user an email verification",
     *     @OA\Response(
     *         response=200,
     *         description="User signed up successfully",
     *         @OA\JsonContent(
     *             @OA\Property(
     *                  property="data", 
     *                  type="object", 
     *                  @OA\Property(
     *                      property="message",
     *                      type="string",
     *                  ),                
     *             ),
     *         ),
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Validation errors"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error"
     *     ),
     *     security={{ "Bearer":{} }}
     *      
     * )
     */
    public function sendEmailVerification(Request $request)
    {
        $user = $request->user();

        if ($user->hasVerifiedEmail()) {
            return response()->json(['data' => ['message' => 'Email is already verified']], 200);
        }

        $user->sendEmailVerificationNotification();
        return response()->json(['data' => ['message' => 'Email verification sent']], 200);
    }

    public function verifyEmail(Request $request)
    {
        $user = User::findOrFail($request->id);

        if ($user->hasVerifiedEmail()) {
            return response()->json(['data' => ['message' => 'Email is already verified']], 200);
        }

        if ($user->markEmailAsVerified()) {
            event(new Verified($user));
        }

        return response()->json(['data' => ['message' => 'Email has been verified']], 200);
    }


    /**
     * @OA\Post(
     *     path="/api/auth/sign-up",
     *     tags={"Authentication"},
     *     summary="User sign-up",
     *     description="Sign up a new user",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(ref="#/components/schemas/SignUpRequest")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="User signed up successfully",
     *         @OA\JsonContent(ref="#/components/schemas/SignUpResponse")
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Validation errors"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error"
     *     )
     * )
     */
    public function signUp(SignUpRequest $request)
    {
        $validatedData = $request->validated();

        $user = User::create([
            'first_name' => $validatedData['first_name'],
            'last_name' => $validatedData['last_name'],
            'email' => $validatedData['email'],
            'password' => bcrypt($validatedData['password']),
        ]);

        $token = $user->createToken($user->email);
        $user->sendEmailVerificationNotification();

        $data = [
            'data' => [
                'message' => 'Registered successfully. Please check your email for verification.',
                'user' => $user,
                'token' => $token->plainTextToken,
            ]
        ];

        return response()->json($data, 200);
    }


    /**
     * @OA\Post(
     *     path="/api/auth/sign-in",
     *     tags={"Authentication"},
     *     summary="User sign-in",
     *     description="Sign in a user",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(ref="#/components/schemas/SignInRequest")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="User signed in successfully",
     *         @OA\JsonContent(ref="#/components/schemas/SignInResponse")
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Validation errors"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error"
     *     )
     * )
     */
    public function signIn(SignInRequest $request)
    {
        $validatedData = $request->validated();

        $user = User::where('email', $validatedData['email'])->first();

        if (!$user || !Hash::check($validatedData['password'], $user->password)) {
            return [
                'message' => 'The provided credentials are incorrect',
                'errors' => []
            ];
        }

        $token = $user->createToken($user->email);

        $data = [
            'data' => [
                'user' => $user,
                'token' => $token->plainTextToken,
            ]
        ];

        return response()->json($data, 200);
    }

    /**
     * @OA\Post(
     *     path="/api/auth/sign-out",
     *     tags={"Authentication"},
     *     summary="Sign out a user",
     *     description="Signs out the authenticated user by invalidating their token.",
     *     @OA\Response(
     *         response=200,
     *         description="User signed out successfully",
     *         @OA\JsonContent(ref="#/components/schemas/SignOutResponse")
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Unauthenticated.")
     *         )
     *     ),
     *     security={{"Bearer": {}}},
     * )
     */
    public function signOut(Request $request)
    {
        $user = $request->user();
        $user->tokens()->delete();

        $data = ['data' => $user];

        return response()->json($data, 200);
    }
}